ath9k: Add u-space control of Virtual and Physical CS
This makes it possible to disable/enable the
virtual and physical carrier sense on the ath9k
wireless cards. To do so, it is needed to mount
debugfs. The appropriate control files are
ieee80211/phy0/ath9k/physical_carrier_sense
ieee80211/phy0/ath9k/virtual_carrier_sense
Writing of "1" disables the CS, whereas "0"
enables it (back to normal operation).

The code is based on:
http://git.popipo.fr/?p=rt2x00.git;a=commit;h=6e617aef04e78aa09fdaf029ff2095bda8509606

Signed-off-by: Rostislav Lisovy <rostislav.lisovy@fel.cvut.cz>