mac80211: Add OCB mode + join and leave handling
All the devices compliant to the IEEE 802.11p
(dot11OCBActivated=true) should operate in the OCB
(Outside the Context of a BSS) mode. This makes it
possible to communicate without the need for
association/authentication. All participants use the
wildcard BSSID (set to all 1's).

In OCB mode there are no extra configuration options
besides the channel (center frequency and width) used
for the communication. The channel must be known a priori.
This is set when joining the network explicitly with the
'ocb join' command.

Signed-off-by: Rostislav Lisovy <rostislav.lisovy@fel.cvut.cz>