its-g5: Add explicit OCB connection state + Fix housekeeping timer
Until now we were unable to distinguish either we were joined
to an OCB network or not. Add 'state' field to struct
ieee80211_if_ocb that is set in "_join" and "_leave" functions.

Since the timer handler might be called during del_timer_sync,
it would cause execution of the ieee80211_ocb_work that would
once again reenable the timer. Add check if we are still connected
to the OCB network in ieee80211_ocb_work. If we are not just
leave the function.

Signed-off-by: Rostislav Lisovy <rostislav.lisovy@fel.cvut.cz>